Boton PWA en navbar
Cambie el boton que estaba en la App a la barra de navegacion
