+++ src/App.js
@@ -5,32 +5,8 @@ import Inicio from './components/pages/Inicio'
 import Shoppingcart from './components/pages/Shoppingcart'
 import Help from './components/pages/Help'
 import Error from './components/pages/Error'
-import React, { useEffect } from 'react'
 
 const App = () =>  {
-  const [deferredPrompt, setDeferredPrompt] = React.useState(null);
-
-  useEffect(() => {
-    const handler = (e) => {
-      e.preventDefault();
-      setDeferredPrompt(e);
-    };
-
-    window.addEventListener('beforeinstallprompt', handler);
-
-    return () => {
-      window.removeEventListener('beforeinstallprompt', handler);
-    };
-  }, []);
-
-  const handleInstall = () => {
-    if (deferredPrompt) {
-      deferredPrompt.prompt();
-      deferredPrompt.userChoice.then(() => {
-        setDeferredPrompt(null);
-      });
-    }
-  };
   return (
     <div className="App">
       <Router>
@@ -42,7 +18,6 @@ const App = () =>  {
           <Route path='/Error' Component={Error}/> 
         </Routes>
       </Router>
-      <button onClick={handleInstall}>Instalar PWA</button>
     </div>
   );
 }
