+++ src/components/navegation/Navbar.js
@@ -1,9 +1,33 @@
-import React from "react";
+import React, { useEffect } from 'react'
 import { Link } from "react-router-dom";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faHome, faShoppingCart, faCircleInfo, faFlag } from '@fortawesome/free-solid-svg-icons';
+import { faHome, faShoppingCart, faCircleInfo, faFlag, faHashtag } from '@fortawesome/free-solid-svg-icons'; 
 
 const Navbar = () => {
+  const [deferredPrompt, setDeferredPrompt] = React.useState(null);
+
+  useEffect(() => {
+    const handler = (e) => {
+      e.preventDefault();
+      setDeferredPrompt(e);
+    };
+
+    window.addEventListener('beforeinstallprompt', handler);
+
+    return () => {
+      window.removeEventListener('beforeinstallprompt', handler);
+    };
+  }, []);
+
+  const handleInstall = () => {
+    if (deferredPrompt) {
+      deferredPrompt.prompt();
+      deferredPrompt.userChoice.then(() => {
+        setDeferredPrompt(null);
+      });
+    }
+  };
+
   return (
     <nav className="navbar navbar-expand-lg navbar-dark bg-dark">
       <div className="container-fluid">
@@ -24,17 +48,17 @@ const Navbar = () => {
         <div className="collapse navbar-collapse" id="navbarSupportedContent">
           <ul className="navbar-nav">
             <li className="nav-item">
-              <Link className="nav-link" to="/">
+              <Link className="nav-link btn btn-primary" to="/">
                 <FontAwesomeIcon icon={faHome} /> Home
               </Link>
             </li>
             <li className="nav-item">
-              <Link className="nav-link" to="/Shoppingcart">
+              <Link className="nav-link btn btn-primary" to="/Shoppingcart">
                 <FontAwesomeIcon icon={faShoppingCart} /> Shopping Cart
               </Link>
             </li>
             <li className="nav-item">
-              <Link className="nav-link" to="/Help">
+              <Link className="nav-link btn btn-primary" to="/Help">
                 <FontAwesomeIcon icon={faCircleInfo} /> Help
               </Link>
             </li>
@@ -42,12 +66,14 @@ const Navbar = () => {
         </div>
         <ul className="navbar-nav ml-auto">
           <li className="nav-item">
+            <button className="nav-link btn btn-warning" onClick={handleInstall}><FontAwesomeIcon icon={faHashtag} />Instalar PWA</button>
+          </li>
+          <li className="nav-item">
             <Link className="nav-link btn btn-danger" to="/Error">
                 <FontAwesomeIcon icon={faFlag} /> Report or Suggestions
             </Link>
           </li>
         </ul>
-        
       </div>
     </nav>
   );
