+++ src/index.js
@@ -10,7 +10,7 @@ root.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById('root')
+  /* document.getElementById('root') */
 );
 
 //serviceWorkerRegistration.register();
